:white_check_mark: resolved suggested changes
