+++ src/helpers/colorHelper.ts
@@ -1,0 +1,9 @@
+const colors: Record<string, string> = {
+  "A+": "#686563",
+  "A-": "#686563",
+  "B+": "#FFF7F5",
+  "B-": "#FFF7F5",
+  F: "#FF6897",
+};
+
+export default colors;
