+++ src/helpers/dateHelper.ts
@@ -1,0 +1,6 @@
+import moment from "moment";
+
+const specificDate = (time: string) =>
+  moment(time).format("MMM D, YYYY [at] hh.mm A");
+
+export default specificDate;
